+++ src/services/utils/index.js
@@ -4,3 +4,4 @@ export { getVersionFromPackage, capFirstLetter } from './strings';
 export { randomInteger, sample } from './number';
 export { isMobile } from './mobile';
 export { objectPropertyAmount } from './object';
+export { JSONService } from '../storage/jsonService';