+++ Changelog.md
@@ -1,5 +1,9 @@
 ## Changelog
 v 2.49
+  Added Start Page improvements, Start page theme, Menu links improvements,
+  Added JSONService to parse, set arrays and objects
+
+v 2.49
   Added Start Page improvements, Start page theme, Menu links improvements
 
 v 2.48